added new serializer

+++ client/FriendConnect/src/com/friendconnect/model/Friend.java
@@ -18,15 +18,13 @@
 
 package com.friendconnect.model;
 
-
 public class Friend extends FriendConnectUser {
 
-	public Friend() {
-
+	public Friend(){
+		
 	}
-
-	public Friend(long id, String nickname, String firstname, String surname,
-			String statusMessage) {
+	
+	public Friend(int id, String nickname, String firstname, String surname, String statusMessage) {
 		this.id = id;
 		this.emailAddress = nickname;
 		this.firstname = firstname;
